tutorial-zero-downtime-db-migrations/requirements.txt
# Database Connectivity
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Configuration
pyyaml==6.0.1

# Logging
coloredlogs==15.0.1

# Type Hints
typing-extensions==4.9.0

# Development
black==23.12.1
flake8==7.0.0
mypy==1.7.1
